+++ src/marionette.controller.js
@@ -8,8 +8,6 @@ Marionette.Controller = function(options){
   this.triggerMethod = Marionette.triggerMethod;
   this.options = options || {};
 
-  Marionette.addEventBinder(this);
-
   if (_.isFunction(this.initialize)){
     this.initialize(this.options);
   }